refactor: merge ffmpeg & mp4decrypt into one module
